GETP-210 refactor: mail 컴포넌트 포트-어댑터 아키텍쳐 적용
