+++ src/main/java/es/princip/getp/application/mail/port/in/SendMailUseCase.java
@@ -1,0 +1,8 @@
+package es.princip.getp.application.mail.port.in;
+
+import es.princip.getp.application.mail.command.SendMailCommand;
+
+public interface SendMailUseCase {
+
+    void send(final SendMailCommand command);
+}
